Adjust LeanCode.IntegrationTestHelpers dependencies
