+++ LeanCode.CoreLibrary.sln
@@ -173,6 +173,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.CQRS.Validation.Fl
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.CQRS.ScopedTransaction", "src\CQRS\LeanCode.CQRS.ScopedTransaction\LeanCode.CQRS.ScopedTransaction.csproj", "{6D962351-01E4-4A32-85C3-1B5A36B0C11C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.Startup.MicrosoftDI", "src\Core\LeanCode.Startup.MicrosoftDI\LeanCode.Startup.MicrosoftDI.csproj", "{5243E4DB-E668-42C3-88F2-38F6135C5F66}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.Components.Autofac", "src\Core\LeanCode.Components.Autofac\LeanCode.Components.Autofac.csproj", "{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1026,6 +1030,30 @@ Global
 		{6D962351-01E4-4A32-85C3-1B5A36B0C11C}.Release|x64.Build.0 = Release|Any CPU
 		{6D962351-01E4-4A32-85C3-1B5A36B0C11C}.Release|x86.ActiveCfg = Release|Any CPU
 		{6D962351-01E4-4A32-85C3-1B5A36B0C11C}.Release|x86.Build.0 = Release|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Debug|x64.Build.0 = Debug|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Debug|x86.Build.0 = Debug|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Release|x64.ActiveCfg = Release|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Release|x64.Build.0 = Release|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Release|x86.ActiveCfg = Release|Any CPU
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66}.Release|x86.Build.0 = Release|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Debug|x64.Build.0 = Debug|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Debug|x86.Build.0 = Debug|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Release|x64.ActiveCfg = Release|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Release|x64.Build.0 = Release|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Release|x86.ActiveCfg = Release|Any CPU
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{7AFA22EA-B093-4457-ABA2-CFF96DF41350} = {7324BABF-C1F7-430D-BF76-CF9BFA559DC6}
@@ -1110,5 +1138,7 @@ Global
 		{DAE6EAC8-676D-41D8-A661-B679114BE897} = {904D67A9-3E75-4754-B802-ACEB2541510B}
 		{A7328190-F3EE-4311-8328-C83490A806B3} = {904D67A9-3E75-4754-B802-ACEB2541510B}
 		{6D962351-01E4-4A32-85C3-1B5A36B0C11C} = {54A36A22-89A6-4525-AB2D-FBBB643C0CE0}
+		{5243E4DB-E668-42C3-88F2-38F6135C5F66} = {7AFA22EA-B093-4457-ABA2-CFF96DF41350}
+		{47EBE467-DDAC-47D5-8669-F5B45AADF9E6} = {7AFA22EA-B093-4457-ABA2-CFF96DF41350}
 	EndGlobalSection
 EndGlobal
